extract chunk typer creation
