+++ webknossos-datastore/app/com/scalableminds/webknossos/datastore/datareaders/ChunkReader.scala
@@ -3,34 +3,16 @@ package com.scalableminds.webknossos.datastore.datareaders
 import com.scalableminds.util.tools.Fox
 import com.scalableminds.util.tools.Fox.box2Fox
 import com.scalableminds.webknossos.datastore.datavault.VaultPath
-import com.typesafe.scalalogging.LazyLogging
 import net.liftweb.common.{Box, Empty, Failure, Full}
 import net.liftweb.util.Helpers.tryo
-import ucar.ma2.{Array => MultiArray, DataType => MADataType}
+import ucar.ma2.{Array => MultiArray}
 
-import java.io.ByteArrayInputStream
-import javax.imageio.stream.MemoryCacheImageInputStream
 import scala.collection.immutable.NumericRange
 import scala.concurrent.ExecutionContext
-import scala.util.Using
 
-object ChunkReader {
-  def create(header: DatasetHeader): ChunkReader =
-    new ChunkReader(header, createChunkTyper(header))
+class ChunkReader(header: DatasetHeader) {
 
-  def createChunkTyper(header: DatasetHeader): ChunkTyper =
-    header.resolvedDataType match {
-      case ArrayDataType.i1 | ArrayDataType.u1 => new ByteChunkTyper(header)
-      case ArrayDataType.i2 | ArrayDataType.u2 => new ShortChunkTyper(header)
-      case ArrayDataType.i4 | ArrayDataType.u4 => new IntChunkTyper(header)
-      case ArrayDataType.i8 | ArrayDataType.u8 => new LongChunkTyper(header)
-      case ArrayDataType.f4                    => new FloatChunkTyper(header)
-      case ArrayDataType.f8                    => new DoubleChunkTyper(header)
-    }
-}
-
-class ChunkReader(val header: DatasetHeader, val chunkTyper: ChunkTyper) {
-  lazy val chunkSize: Int = header.chunkSize.toList.product
+  private lazy val chunkTyper = ChunkTyper.createFromHeader(header)
 
   def read(path: VaultPath, chunkShapeFromMetadata: Array[Int], range: Option[NumericRange[Long]])(
       implicit ec: ExecutionContext): Fox[MultiArray] =
@@ -58,102 +40,4 @@ class ChunkReader(val header: DatasetHeader, val chunkTyper: ChunkTyper) {
       }
       decompressed <- tryo(header.compressorImpl.decompress(bytes)).toFox ?~> "chunk.decompress.failed"
     } yield (decompressed, None)
-}
-
-abstract class ChunkTyper {
-  val header: DatasetHeader
-
-  def ma2DataType: MADataType
-  def wrapAndType(bytes: Array[Byte], chunkShape: Array[Int]): MultiArray
-
-  def createFromFillValue(chunkShape: Array[Int]): MultiArray =
-    MultiArrayUtils.createFilledArray(ma2DataType, chunkShape, header.fillValueNumber)
-
-  // Chunk shape in header is in C-Order (XYZ), but data may be in F-Order (ZYX), so the chunk shape
-  // associated with the array needs to be adjusted for non-isotropic chunk sizes.
-  def chunkSizeOrdered(chunkSize: Array[Int]): Array[Int] =
-    if (header.order == ArrayOrder.F) chunkSize.reverse else chunkSize
-}
-
-class ByteChunkTyper(val header: DatasetHeader) extends ChunkTyper {
-  val ma2DataType: MADataType = MADataType.BYTE
-
-  def wrapAndType(bytes: Array[Byte], chunkShape: Array[Int]): MultiArray =
-    MultiArray.factory(ma2DataType, chunkSizeOrdered(chunkShape), bytes)
-}
-
-class DoubleChunkTyper(val header: DatasetHeader) extends ChunkTyper {
-
-  val ma2DataType: MADataType = MADataType.DOUBLE
-
-  def wrapAndType(bytes: Array[Byte], chunkShape: Array[Int]): MultiArray =
-    Using.Manager { use =>
-      val typedStorage = new Array[Double](chunkShape.product)
-      val bais = use(new ByteArrayInputStream(bytes))
-      val iis = use(new MemoryCacheImageInputStream(bais))
-      iis.setByteOrder(header.byteOrder)
-      iis.readFully(typedStorage, 0, typedStorage.length)
-      MultiArray.factory(ma2DataType, chunkSizeOrdered(chunkShape), typedStorage)
-    }.get
-}
-
-class ShortChunkTyper(val header: DatasetHeader) extends ChunkTyper with LazyLogging {
-
-  val ma2DataType: MADataType = MADataType.SHORT
-
-  def wrapAndType(bytes: Array[Byte], chunkShape: Array[Int]): MultiArray =
-    Using.Manager { use =>
-      val typedStorage = new Array[Short](chunkShape.product)
-      val bais = use(new ByteArrayInputStream(bytes))
-      val iis = use(new MemoryCacheImageInputStream(bais))
-      iis.setByteOrder(header.byteOrder)
-      iis.readFully(typedStorage, 0, typedStorage.length)
-      MultiArray.factory(ma2DataType, chunkSizeOrdered(chunkShape), typedStorage)
-    }.get
-}
-
-class IntChunkTyper(val header: DatasetHeader) extends ChunkTyper {
-
-  val ma2DataType: MADataType = MADataType.INT
-
-  def wrapAndType(bytes: Array[Byte], chunkShape: Array[Int]): MultiArray =
-    Using.Manager { use =>
-      val typedStorage = new Array[Int](chunkShape.product)
-      val bais = use(new ByteArrayInputStream(bytes))
-      val iis = use(new MemoryCacheImageInputStream(bais))
-      iis.setByteOrder(header.byteOrder)
-      iis.readFully(typedStorage, 0, typedStorage.length)
-      MultiArray.factory(ma2DataType, chunkSizeOrdered(chunkShape), typedStorage)
-    }.get
-}
-
-class LongChunkTyper(val header: DatasetHeader) extends ChunkTyper {
-
-  val ma2DataType: MADataType = MADataType.LONG
-
-  def wrapAndType(bytes: Array[Byte], chunkShape: Array[Int]): MultiArray =
-    Using.Manager { use =>
-      val typedStorage = new Array[Long](chunkShape.product)
-      val bais = use(new ByteArrayInputStream(bytes))
-      val iis = use(new MemoryCacheImageInputStream(bais))
-      iis.setByteOrder(header.byteOrder)
-      iis.readFully(typedStorage, 0, typedStorage.length)
-      MultiArray.factory(ma2DataType, chunkSizeOrdered(chunkShape), typedStorage)
-    }.get
-
-}
-
-class FloatChunkTyper(val header: DatasetHeader) extends ChunkTyper {
-
-  val ma2DataType: MADataType = MADataType.FLOAT
-
-  def wrapAndType(bytes: Array[Byte], chunkShape: Array[Int]): MultiArray =
-    Using.Manager { use =>
-      val typedStorage = new Array[Float](chunkShape.product)
-      val bais = use(new ByteArrayInputStream(bytes))
-      val iis = use(new MemoryCacheImageInputStream(bais))
-      iis.setByteOrder(header.byteOrder)
-      iis.readFully(typedStorage, 0, typedStorage.length)
-      MultiArray.factory(ma2DataType, chunkSizeOrdered(chunkShape), typedStorage)
-    }.get
 }
