+++ webknossos-datastore/app/com/scalableminds/webknossos/datastore/datareaders/n5/N5ChunkReader.scala
@@ -2,7 +2,7 @@ package com.scalableminds.webknossos.datastore.datareaders.n5
 
 import com.scalableminds.util.tools.Fox
 import com.scalableminds.util.tools.Fox.box2Fox
-import com.scalableminds.webknossos.datastore.datareaders.{ChunkReader, ChunkTyper, DatasetHeader}
+import com.scalableminds.webknossos.datastore.datareaders.{ChunkReader, DatasetHeader}
 import com.scalableminds.webknossos.datastore.datavault.VaultPath
 import com.typesafe.scalalogging.LazyLogging
 import net.liftweb.util.Helpers.tryo
@@ -10,19 +10,12 @@ import net.liftweb.util.Helpers.tryo
 import scala.collection.immutable.NumericRange
 import scala.concurrent.ExecutionContext
 
-object N5ChunkReader {
-  def create(header: DatasetHeader): ChunkReader =
-    new N5ChunkReader(header, ChunkReader.createChunkTyper(header))
-}
-
 // N5 allows for a 'varmode' which means that the number of elements in the chunk can deviate from the set chunk size.
 // This is used for end chunks, but apparently also for overlapping (see https://github.com/zarr-developers/zarr-specs/issues/44#issuecomment-360669270).
 // Here, we provide only provide one implementation to handle the `varmode`:
 // N5ChunkReader, always fills the chunk to the bytes necessary
 
-class N5ChunkReader(header: DatasetHeader, typedChunkReader: ChunkTyper)
-    extends ChunkReader(header, typedChunkReader)
-    with LazyLogging {
+class N5ChunkReader(header: DatasetHeader) extends ChunkReader(header) with LazyLogging {
 
   private val dataExtractor: N5DataExtractor = new N5DataExtractor
